% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{network_genes}
\alias{network_genes}
\title{List all gene symbols in the network}
\usage{
network_genes()
}
\value{
Character vector of gene symbols (union of both T matrices).
}
\description{
List all gene symbols in the network
}
